requirements.txt
numpy
pandas
pulp
typing
Levenshtein
matplotlib
pyyaml